requirements.txt
python-telegram-bot==11.1.0
django==1.11.7
psycopg2==2.7.6.1